requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
feedparser
azure-data-tables
azure-communication-email

python-dotenv
requests
jsons